Mini V3 Dev Kit/requirements.txt
scipy==1.8.1
numpy>=1.23.2
opencv-python>=4.0.0
flask>=2.2.2
pyserial>=3.5
